test(cron-leases): extract the shared lease fixture into a test-support module
The cron-lease suites carried one byte-identical 229-line block: the
sqlite-backed D1 lease fixture, its five hand-written row types, and the
TestLeaseDb accessor surface. cron-leases.test.ts:25-250 and
cron-leases-scheduled-slot.test.ts:10-249 were the same text, so any
schema change had to be applied twice or the two suites would silently
disagree about the shape of the tables they assert on.

Move the fixture to cron-leases.test-support.ts, following the existing
*.test-support.ts convention already used by status, telegram-webhook,
dispatch-telegram-alerts, enrich-prices and sync-yield-data. The module
exports makeLeaseDb, setSlotUpdatedAt and closeOpenLeaseDatabases; the
row types stay unexported because neither suite names them, keeping the
new surface to the three symbols that are actually imported.

Scaffolding only - no assertion changed. Both suites keep their exact
it() counts (27 and 26) and the 53 tests pass unchanged. The lease
suite's afterEach now calls closeOpenLeaseDatabases() in place of the
inlined splice-and-close loop; the slot suite's teardown is deliberately
left as it was, so behaviour is identical on both sides.

Net 217 lines removed. This is the reference case for the test-support
consolidation wave documented in agents/2026-08-23-recon/PLAN.md, which
measured 17,794 exactly-duplicated lines across the repo with 84% of
them in test files.

Follow-up, intentionally not bundled: the slot suite never closed its
sqlite handles, so every makeLeaseDb() there leaks one. That is a
one-line resource fix and belongs in its own commit rather than inside a
consolidation that must not change behaviour.
